+++ packages/button/src/__tests__/Button.screenshots.test.tsx
@@ -1,9 +1,8 @@
 import {
-  closeBrowser,
   createStorybookUrl,
-  matchHtml,
-  openBrowserPage,
   setupScreenshotTesting,
+  screenshotMatchHover,
+  screenshotMatchClick,
 } from '../../../screenshotUtils';
 
 const packageName = 'button';
@@ -80,50 +79,14 @@ describe('Button events tests', () => {
       packageName,
     });
 
-    const { browser, context, page } = await openBrowserPage(pageUrl);
-
-    try {
-      await page.hover(buttonSelector);
-
-      await matchHtml({
-        page,
-        expect,
-        screenshotOpts: { clip },
-      });
-    } catch (error: unknown) {
-      console.error(error);
-    } finally {
-      await closeBrowser({ browser, context, page });
-    }
+    await screenshotMatchHover(pageUrl, buttonSelector, { clip });
   });
 
   test('Button click test', async () => {
     const pageUrl = createStorybookUrl({
       packageName,
     });
-    const { browser, context, page } = await openBrowserPage(pageUrl);
 
-    try {
-      const buttons = await page.$$(buttonSelector);
-
-      buttons.forEach(async button => {
-        const buttonPosition = await button.boundingBox();
-
-        if (buttonPosition) {
-          await page.mouse.move(buttonPosition.x, buttonPosition.y);
-          await page.mouse.down();
-        }
-      });
-
-      await matchHtml({
-        page,
-        expect,
-        screenshotOpts: { clip },
-      });
-    } catch (error: unknown) {
-      console.error(error);
-    } finally {
-      await closeBrowser({ browser, context, page });
-    }
+    await screenshotMatchClick(pageUrl, buttonSelector, { clip });
   });
 });