+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# [1.9.0](https://github.com/hey-car/heycar-uikit/compare/v1.8.0...v1.9.0) (2022-08-03)
+
+
+### Features
+
+* **badge:** The first implantation of badge component ([#42](https://github.com/hey-car/heycar-uikit/issues/42)) ([2ca5a7e](https://github.com/hey-car/heycar-uikit/commit/2ca5a7ea897c4cdeccaadca4f4d1592be40fe6e7))
+
+# [1.8.0](https://github.com/hey-car/heycar-uikit/compare/v1.7.0...v1.8.0) (2022-08-02)
+
+
+### Features
+
+* **icons:** updated icons ([#45](https://github.com/hey-car/heycar-uikit/issues/45)) ([2293b7a](https://github.com/hey-car/heycar-uikit/commit/2293b7ad0135fb645495afb9a11583ce8ef7e6e3))
+
+# [1.7.0](https://github.com/hey-car/heycar-uikit/compare/v1.6.0...v1.7.0) (2022-07-27)
+
+
+### Features
+
+* **switch:** initial implementation of the switch component ([#41](https://github.com/hey-car/heycar-uikit/issues/41)) ([fae4001](https://github.com/hey-car/heycar-uikit/commit/fae40014e387fd6a84275fdee825b265b798b7d3))
+
+# [1.6.0](https://github.com/hey-car/heycar-uikit/compare/v1.5.0...v1.6.0) (2022-07-25)
+
+
+### Features
+
+* **icons:** the first implementation of the icons package ([#39](https://github.com/hey-car/heycar-uikit/issues/39)) ([74fd00d](https://github.com/hey-car/heycar-uikit/commit/74fd00d9227e3e915fbb54962d56ab9d64715b5c))
+
+# [1.5.0](https://github.com/hey-car/heycar-uikit/compare/v1.4.0...v1.5.0) (2022-07-22)
+
+
+### Features
+
+* **icons:** the first implementation of the icons package ([#40](https://github.com/hey-car/heycar-uikit/issues/40)) ([60e2055](https://github.com/hey-car/heycar-uikit/commit/60e205578ce2d2024eb22c0a8384a17d7fbe4b91))
+
 # [1.4.0](https://github.com/hey-car/heycar-uikit/compare/v1.3.0...v1.4.0) (2022-07-14)