+++ packages/themes/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/themes@1.1.0...@heycar-uikit/themes@1.1.0) (2022-08-03)
+
+
+### Features
+
+* **badge:** The first implantation of badge component ([#42](https://github.com/hey-car/heycar-uikit/issues/42)) ([2ca5a7e](https://github.com/hey-car/heycar-uikit/commit/2ca5a7ea897c4cdeccaadca4f4d1592be40fe6e7))
+* **icons:** the first implementation of the icons package ([#39](https://github.com/hey-car/heycar-uikit/issues/39)) ([74fd00d](https://github.com/hey-car/heycar-uikit/commit/74fd00d9227e3e915fbb54962d56ab9d64715b5c))
+* **icons:** the first implementation of the icons package ([#40](https://github.com/hey-car/heycar-uikit/issues/40)) ([60e2055](https://github.com/hey-car/heycar-uikit/commit/60e205578ce2d2024eb22c0a8384a17d7fbe4b91))
+* **switch:** initial implementation of the switch component ([#41](https://github.com/hey-car/heycar-uikit/issues/41)) ([fae4001](https://github.com/hey-car/heycar-uikit/commit/fae40014e387fd6a84275fdee825b265b798b7d3))
+
+
+
+
+
+# [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/themes@1.1.0...@heycar-uikit/themes@1.1.0) (2022-07-27)
+
+
+### Features
+
+* **icons:** the first implementation of the icons package ([#39](https://github.com/hey-car/heycar-uikit/issues/39)) ([74fd00d](https://github.com/hey-car/heycar-uikit/commit/74fd00d9227e3e915fbb54962d56ab9d64715b5c))
+* **icons:** the first implementation of the icons package ([#40](https://github.com/hey-car/heycar-uikit/issues/40)) ([60e2055](https://github.com/hey-car/heycar-uikit/commit/60e205578ce2d2024eb22c0a8384a17d7fbe4b91))
+* **switch:** initial implementation of the switch component ([#41](https://github.com/hey-car/heycar-uikit/issues/41)) ([fae4001](https://github.com/hey-car/heycar-uikit/commit/fae40014e387fd6a84275fdee825b265b798b7d3))
+
+
+
+
+
+# [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/themes@1.1.0...@heycar-uikit/themes@1.1.0) (2022-07-25)
+
+
+### Features
+
+* **icons:** the first implementation of the icons package ([#39](https://github.com/hey-car/heycar-uikit/issues/39)) ([74fd00d](https://github.com/hey-car/heycar-uikit/commit/74fd00d9227e3e915fbb54962d56ab9d64715b5c))
+* **icons:** the first implementation of the icons package ([#40](https://github.com/hey-car/heycar-uikit/issues/40)) ([60e2055](https://github.com/hey-car/heycar-uikit/commit/60e205578ce2d2024eb22c0a8384a17d7fbe4b91))
+
+
+
+
+
+# [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/themes@1.1.0...@heycar-uikit/themes@1.1.0) (2022-07-22)
+
+
+### Features
+
+* **icons:** the first implementation of the icons package ([#40](https://github.com/hey-car/heycar-uikit/issues/40)) ([60e2055](https://github.com/hey-car/heycar-uikit/commit/60e205578ce2d2024eb22c0a8384a17d7fbe4b91))
+
+
+
+
+
 # [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/themes@1.0.3...@heycar-uikit/themes@1.1.0) (2022-07-14)